// frontend/src/app/model/Produit.ts
export class Produit {
    
    prod_numero: number;
    prod_numerobranche: number;
    prod_numerocategorie: number;
    prod_denominationlong: String ;
    prod_denominationcourt: String ;
    prod_numerogarantieassoc1: number ;
    prod_numerogarantieassoc2: number ;
    prod_numerogarantieassoc3: number ;
	prod_numerogarantieassoc4: number ;
    prod_numerogarantieassoc5: number ;
    prod_numerogarantieassoc6: number ;
    prod_numerogarantieassoc7: number ;
    prod_numerogarantieassoc8: number ;
    prod_numerogarantieassoc9: number ;
    prod_numerogarantieassoc10: number ;
    prod_numeroextension1: number ;
    prod_numeroextension2: number ;
    prod_numeroextension3: number ;
	prod_codetaxe: number ;
    prod_remisecommerciale: number ;
	prod_remiseexceptionnelle: number ;
    prod_codeutilisateur: String ;
    prod_datemodification: Date ;
}